upgrade: clean code

+++ kharazim-application/src/main/java/com/tyrael/kharazim/application/base/BasePageMapper.java
@@ -21,7 +21,7 @@ public interface BasePageMapper<T> extends BaseMapper<T> {
      */
     default PageResponse<T> page(Page<T> pageCondition, Wrapper<T> queryWrapper) {
         Page<T> pageData = selectPage(pageCondition, queryWrapper);
-        return PageResponse.success(pageData.getRecords(), pageData.getTotal());
+        return PageResponse.success(pageData.getRecords(), (int) pageData.getTotal());
     }
 
     /**
